requirements.txt
Django==2.2.6
djangorestframework==3.10.3
psycopg2-binary
gunicorn
drf-yasg==1.17.0
swagger_spec_validator
